# server/openapi/api.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Receipt Processor
 *
 * A simple receipt processor
 *
 * API version: 1.0.0
 */

package openapi

import (
	"context"
	"log"
	"net/http"
)

// DefaultAPIRouter defines the required methods for binding the api requests to a responses for the DefaultAPI
// The DefaultAPIRouter implementation should parse necessary information from the http request,
// pass the data to a DefaultAPIServicer to perform the required actions, then write the service results to the http response.
type DefaultAPIRouter interface { 
	ProcessReceipt(http.ResponseWriter, *http.Request)
	GetReceiptPoints(http.ResponseWriter, *http.Request)
}


// DefaultAPIServicer defines the api actions for the DefaultAPI service
// This interface intended to stay up to date with the openapi yaml used to generate it,
// while the service implementation can be ignored with the .openapi-generator-ignore file
// and updated with the logic required for the API.
type DefaultAPIServicer interface { 
	ProcessReceipt(context.Context, Receipt) (ImplResponse, error)
	GetReceiptPoints(context.Context, string) (ImplResponse, error)
}

type MyService struct {}

func NewMyService() *MyService {
	return &MyService{}
}

func (s *MyService) ProcessReceipt(ctx context.Context, receipt Receipt) (ImplResponse, error) {
	// thing, err := s.store.GetThing(uuid)
	// if err == helpers.ErrNotFound {
	// 	return openapi.Response(http.StatusNotFound, nil), nil
	// }
	// if err != nil {
	// 	return openapi.Response(http.StatusInternalServerError, nil), err
	// }
	log.Println("Got request to process receipt")
	return Response(http.StatusOK, "Heyo!"), nil
}

func (s *MyService) GetReceiptPoints(context.Context, string) (ImplResponse, error) {
	log.Println("Got request to get receipt points")
	return Response(http.StatusOK, "Heyo!"), nil
}