# server/openapi/model_item.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Receipt Processor
 *
 * A simple receipt processor
 *
 * API version: 1.0.0
 */

package openapi




type Item struct {

	// The Short Product Description for the item.
	ShortDescription string `json:"shortDescription" validate:"regexp=^[\\\\w\\\\s\\\\-]+$"`

	// The total price payed for this item.
	Price string `json:"price" validate:"regexp=^\\\\d+\\\\.\\\\d{2}$"`
}

// AssertItemRequired checks if the required fields are not zero-ed
func AssertItemRequired(obj Item) error {
	elements := map[string]interface{}{
		"shortDescription": obj.ShortDescription,
		"price": obj.Price,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertItemConstraints checks if the values respects the defined constraints
func AssertItemConstraints(obj Item) error {
	return nil
}